Add files via upload
For me as a beginner, I had to look into why the notation flipped between U_f and U_\omega for the oracle. The "Alternative oracle definition" section of https://en.wikipedia.org/wiki/Grover%27s_algorithm#Problem_description makes me think the U_f is incorrect because there isn't an ancillary bit and the reflection is directly on the state being checked. Also most of the page on Grover's has U_\omega.